+++ draw.hpp
@@ -29,12 +29,12 @@ void initGlfw(const char * title, const int screenW, const int screenH);
 void initGfx();
 
 
-void drawTree(sSquare* tree, const size_t W, const size_t H, const float dX, const float dY);
+void drawTree(sSquare* tree, const uint32_t W, const uint32_t H, const float dX, const float dY);
 ///////////
 // draw game content
 void drawPlanets(const saPlanet & sPlanets, const float dX, const float dY);
 void drawShips(const saShip * ships, const float dX, const float dY);
-void drawShipMarkers(const sShip * const ships, std::vector<size_t> & selectedShips, const float dX, const float dY);
+void drawShipMarkers(const sShip * const ships, std::vector<uint32_t> & selectedShips, const float dX, const float dY);
 void drawShots(const saShot * shots, const float dX, const float dY);
 
 void drawRectangle(vec2 v1, vec2 v2, const float dX, const float dY);
@@ -42,8 +42,8 @@ void drawRectangle(vec2 v1, vec2 v2, const float dX, const float dY);
 // draw text/numbers/strings
 void drawInt(int i, float strX, float strY, float stretchX, float stretchY);
 void drawInt(int i, float strX, float strY, float stretchXY=1);
-void drawString(const char* str, size_t strlen, float strX, float strY, float stretchX, float stretchY);
-void drawString(const char* str, size_t strlen, float strX, float strY, float stretchXY=1);
+void drawString(const char* str, uint32_t strlen, float strX, float strY, float stretchX, float stretchY);
+void drawString(const char* str, uint32_t strlen, float strX, float strY, float stretchXY=1);
 
 ////////////////////
 //  other stuff
