+++ game.hpp
@@ -12,13 +12,14 @@
 class Game {  
   public:
     struct GameConfig {
-      size_t numPlanets;
-      size_t numShips;
-      size_t numShots;
+      /// networked, use constsize types only
+      uint32_t numPlanets;
+      uint32_t numShips;
+      uint32_t numShots;
       vec2 map;
       float money[2];
     };
-    static GameConfig createConfig(size_t const NUM_PLANETS, size_t const NUM_SHIPS, vec2 const map, float moneyA=0, float moneyB=0);
+    static GameConfig createConfig(uint32_t const NUM_PLANETS, uint32_t const NUM_SHIPS, vec2 const map, float moneyA=0, float moneyB=0);
   public:
     float mMoney[PN]; // money of PA and PB
     vec2 mMap;
@@ -27,16 +28,16 @@ class Game {
     saShot mShots[2];
     saShip mShips[2];
     
-    size_t treeW = 0;
-    size_t treeH = 0;
+    uint32_t treeW = 0;
+    uint32_t treeH = 0;
     sSquare *mTree = nullptr;
   
-    std::vector<size_t> selectedShips;
+    std::vector<uint32_t> selectedShips;
   private:
     void GameCtor(GameConfig cfg);
     GameConfig config;
   public:
-    Game(vec2 map, const size_t MAX_SHIPS, const size_t NUM_PLANETS);
+    Game(vec2 map, const uint32_t MAX_SHIPS, const uint32_t NUM_PLANETS);
     Game(GameConfig cfg);
     ~Game();
     
@@ -45,16 +46,16 @@ class Game {
     void update(const double dt, const bool bUpdatePlanets = true);
     void shootAndCollide();
     
-    void * packData(size_t & size, double time); // 
-    double unpackData(void * const data, size_t size, const double time); // returns timeDelta
+    void * packData(uint32_t & size, double time); // 
+    double unpackData(void * const data, uint32_t size, const double time); // returns timeDelta
 
-    void * packUpdateData(size_t & size, double time); // 
-    double unpackUpdateData(void * const data, size_t size, const double time); // returns timeDelta
+    void * packUpdateData(uint32_t & size, double time); // 
+    double unpackUpdateData(void * const data, uint32_t size, const double time); // returns timeDelta
     
     void select(Party party, vec2 v);
     void select(Party party, vec2 v1, vec2 v2);
     /// command all selected ships to go somewhere
-    void * sendSelectedGetData(Party party, vec2 v1, vec2 v2, size_t formation, size_t & size);
+    void * sendSelectedGetData(Party party, vec2 v1, vec2 v2, uint32_t formation, uint32_t & size);
     void sendShips(Party party, void * const data);
     
     void clearChanged();
@@ -65,15 +66,15 @@ class Game {
     
   private:
     /// init list of game objects
-    void initGame(saPlanet & planets, saShip * ships, saShot * shots, const size_t MAX_SHIPS);
-    void initPlanets(saPlanet & planets, const size_t size);
-    void initShips  (saShip   & ships,   const size_t size);
-    void initShots  (saShot   & shots,   const size_t size);
+    void initGame(saPlanet & planets, saShip * ships, saShot * shots, const uint32_t MAX_SHIPS);
+    void initPlanets(saPlanet & planets, const uint32_t size);
+    void initShips  (saShip   & ships,   const uint32_t size);
+    void initShots  (saShot   & shots,   const uint32_t size);
     /// process list of game objects
     void updatePlanets(const double dt);
     void updateShips(const double dt);
     void updateShots(const double dt);
-    void updateShotsIntervall(saShot & sShots, const size_t pStart, const size_t num, const double dt); 
+    void updateShotsIntervall(saShot & sShots, const uint32_t pStart, const uint32_t num, const double dt); 
     void updateShip(sShip & ship, const double dt); // single ship update
     /**
     ship/planet - ship/planet to check for closest enemy and shoot
@@ -83,33 +84,33 @@ class Game {
     W,H - Size of rivalTree
     */
     // todo split into shoot and collision detection, also separate the drawTree call from here 
-    bool shoot(sShip & ship, saPlanet & sPlanets, saShot & shots, sSquare * rivalTree, const size_t W, const size_t H, Party party);
-    bool shoot(sPlanet & planet, saPlanet & sPlanets, saShot & shots, sSquare * rivalTree, const size_t W, const size_t H);
+    bool shoot(sShip & ship, saPlanet & sPlanets, saShot & shots, sSquare * rivalTree, const uint32_t W, const uint32_t H, Party party);
+    bool shoot(sPlanet & planet, saPlanet & sPlanets, saShot & shots, sSquare * rivalTree, const uint32_t W, const uint32_t H);
     /// helping functions
     inline double vecLen(const vec2 v);
     inline void delta(const float x, const float y, const float tx, const float ty, float & dx, float & dy); /// dx=tx-x
     inline void normalize(float & x, float & y, const float LEN);  /// normalized vector * LEN
-    inline size_t distanceSQ(const float x, const float y, const float x2, const float y2); /// distance^2
+    inline uint32_t distanceSQ(const float x, const float y, const float x2, const float y2); /// distance^2
     /// command ship/shot to go somewhere
     void flyToTarget(sShot & shot, const float tx, const float ty);
-    void flyToTarget(Party party, const size_t id, const float tx, const float ty);
+    void flyToTarget(Party party, const uint32_t id, const float tx, const float ty);
     /// insert or delete ships or shots
     bool addShot(Party party, const float x, const float y, const float tx, const float ty);
     bool addShip(Party party, const float x, const float y, const float tx, const float  ty);
-    void deleteShip(Party party, const size_t id);
+    void deleteShip(Party party, const uint32_t id);
     /// damage ships or planets
     void takeDamage(sShip & ship);
-    void takeDamage(sPlanet & planet, const size_t party);
-    void capturePlanet(sPlanet & planet, const size_t newParty);
+    void takeDamage(sPlanet & planet, const uint32_t party);
+    void capturePlanet(sPlanet & planet, const uint32_t newParty);
     /// upgrade planets
     bool upgradePlanet(sPlanet & planet, Upgrade upgrade);
     
     
     
     /// un/packing gameData
-    void * packChangedShips(Party party, size_t & size);
+    void * packChangedShips(Party party, uint32_t & size);
     void unpackChangedShips(Party party, void * const data, const double dt);
-    void * packChangedShots(Party party, size_t & size);
+    void * packChangedShots(Party party, uint32_t & size);
     void unpackChangedShots(Party party, void * const data, const double dt);
 };
