+++ net.hpp
@@ -19,14 +19,14 @@ enum PacketType {
   PTYPE_PAUSE,              // S pauses game
 };
 
-ENetPacket* enet_packet_create(	const void * 	data, size_t 	dataLength, enet_uint32 	flags, const PacketType packetType);
+ENetPacket* enet_packet_create(	const void * data, uint32_t 	dataLength, enet_uint32 flags, const PacketType packetType);
 enet_uint8 * enet_packet_data(ENetPacket * packet);
 PacketType enet_packet_type(ENetPacket * packet);
-size_t enet_packet_size(ENetPacket * packet);
+uint32_t enet_packet_size(ENetPacket * packet);
 
 
 /// creates an enet packet with a packetType added in the beginning
-ENetPacket* enet_packet_create(	const void * 	data, size_t 	dataLength, enet_uint32 	flags, PacketType packetType) {
+ENetPacket* enet_packet_create(	const void * 	data, uint32_t 	dataLength, enet_uint32 	flags, PacketType packetType) {
   ENetPacket* packet = enet_packet_create(nullptr, sizeof(__PTYPE)+dataLength, flags);
   *(__PTYPE*)packet->data = (__PTYPE)packetType;
   memcpy(packet->data+sizeof(__PTYPE), data, dataLength);
@@ -41,7 +41,7 @@ enet_uint8 * enet_packet_data(ENetPacket * packet){
   return packet->data+sizeof(__PTYPE);
 }
 /// gets the size of the data, packettype-exclusive
-size_t enet_packet_size(ENetPacket * packet){
+uint32_t enet_packet_size(ENetPacket * packet){
   return packet->dataLength - sizeof(__PTYPE);
 }
